pyproject.toml
[tool.poetry]
name = "новая папка 2"
version = "0.1.0"
description = ""
authors = ["user <panpanzov.nur@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.7"
psycopg2-binary = "^2.9.7"
pillow = "^10.0.1"
httpx = "^0.25.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
